+++ providers/capvcd/standard/capvcd_test.go
@@ -6,11 +6,11 @@ import (
 	"github.com/giantswarm/cluster-test-suites/internal/common"
 )
 
-var _ = PDescribe("Common tests", func() {
+var _ = Describe("Common tests", func() {
 	common.Run(&common.TestConfig{
 		// No autoscaling on-prem
 		AutoScalingSupported: false,
-		BastionSupported:     true,
+		BastionSupported:     false,
 		// Disabled until https://github.com/giantswarm/roadmap/issues/1037
 		ExternalDnsSupported: false,
 	})