+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Re-enabled CAPVCD tests
 - Small cleanup tasks - added some relevant comments, explicitly set false bools, added logging to scale test
 
 ## [1.17.0] - 2023-11-15